+++ changelog.md
@@ -1,6 +1,21 @@
 1.5.2
-- Added socket "GETCARDSBYDECK" that returns JournalEntries of all cards in a given deck. It does this by calling the hool
-`deckID-info` with (deckID, cards) where cards is a list of journal entries in the deck
+- Added socket "GETCARDSBYDECK" that returns JournalEntries of all cards in a given deck. It does this by returning a socket message: 
+
+You send it this
+export interface MSG_GETALLCARDSBYDECK {
+  type: "GETALLCARDSBYDECK",
+  playerID: string, //GM USER ID
+  to: string, //PLAYER USER ID THAT YOU WANT THE CARDS SENT TO
+  deckID: string
+}
+
+It returns you back:
+export interface MSG_RECEIVECARDSBYDECK {
+  type: "RECEIVECARDSBYDECK",
+  playerID: string,
+  cards: JournalEntry[],
+  deckID: string
+}
 
 1.5.1
 - Fixed the deck not coming onto the scene (needed to add legacy support for deck images not being on older decks)